Adding support for custom complex objects.
After looking at the output generated from a more complex report it has led us to the conclusion that using a format of json and then using the xsds for schema generation will be ideal. Using format=json makes syncing the data easy and the xsds includes sufficient type information for the JSON to allow the transformer to properly cast decimals.

Part of this change includes manipulating the URL to force it to have a format=json query parameter. This will allow the tap to work without needing our customers to update their config.