Support for complex nested objects (#2)
* wip: discovery working untested for complex types

* wip: synced records pass transformation

* Change report in integration test to one containing complex types

* Remove unused variable

* update test start_date

* Change test report to new one with complex types

* Check whether the element is an array

* WIP initial unit tests around discovery

* WIP actually add the test

* force usage of the JSON format

* Updated comment

* Remove log line

* Fix inconsistent discovery URL

* Handle the json iteratively

* Defensive code updates

* Adding support for custom complex objects.

After looking at the output generated from a more complex report it has led us to the conclusion that using a format of json and then using the xsds for schema generation will be ideal. Using format=json makes syncing the data easy and the xsds includes sufficient type information for the JSON to allow the transformer to properly cast decimals.

Part of this change includes manipulating the URL to force it to have a format=json query parameter. This will allow the tap to work without needing our customers to update their config.

* Fix copypasta

* Include bad value in Exception

* Remove the format: decimal until that is available

* Fix typo and do the change to both exceptions

Co-authored-by: Collin Simon <csimon@talend.com>